fix(build): mover helpers lojistas-credito pra lib/ (Next proíbe exports arbitrários em route.ts)
